refactor: ImageType을 entity에서 enums 패키지로 이동
  ImageType을 엔티티에서 열거형으로 적절한 위치로 재배치하여 도메인 구조 개선
