+++ src/main/java/com/divary/domain/image/dto/response/ImageResponse.java
@@ -1,7 +1,8 @@
 package com.divary.domain.image.dto.response;
 
 import com.divary.domain.image.entity.Image;
-import com.divary.domain.image.entity.ImageType;
+import com.divary.domain.image.enums.ImageType;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
