+++ src/main/java/com/divary/domain/image/enums/ImageType.java
@@ -1,4 +1,4 @@
-package com.divary.domain.image.entity;
+package com.divary.domain.image.enums;
 
 import lombok.Getter;
